Add files via upload
The wishlist is now a static class, the list itself is private.  This means that the same wishlist is accessible everywhere in the program.
A method has been added to the Wishlist.UI designer code that brings the wishlist panel to the front, and calls the display wishlist function.  This fixed an error where we could not call display wishlist from Form1.cs because the listbox on Wishlist.UI was private.